0018: A
Changing where the `long` is because surely that affects something

+++ 0018.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    vector<vector<int>> fourSum(vector<int>& nums, int target) {
+    vector<vector<int>> fourSum(vector<int>& nums, long target) {
         int n = nums.size();
         set<vector<int>> ans;
         unordered_map<long, set<vector<int>>> umap;
@@ -15,7 +15,7 @@ public:
             }
             int c = b + 1;
             for (int d = b + 2; d < n; d++) {
-                long rhs = (long)target - nums[c] - nums[d];
+                long rhs = target - nums[c] - nums[d];
                 for (auto v : umap[rhs]) {
                     v.push_back(nums[c]);
                     v.push_back(nums[d]);
